+++ src/Http/Requests/Dominion/Actions/GovernmentActionRequest.php
@@ -6,4 +6,16 @@ use OpenDominion\Http\Requests\Dominion\AbstractDominionRequest;
 
 class GovernmentActionRequest extends AbstractDominionRequest
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function rules()
+    {
+        $rules = [
+            'appointee' => 'required|integer|exists:dominions,id',
+            'role' => 'required|string'
+        ];
+
+        return $rules;
+    }
 }